requirements.txt
# requirements.txt
streamlit
langchain
langchain-groq
langgraph
chromadb
transformers
torch
python-dotenv
pydantic